pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["thiagodiasv <thiago76ers@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8, <3.9"
django = "^3.1"
djangorestframework = "^3.11.1"
drf-yasg = "^1.17.1"
djangorestframework_simplejwt = "^4.4.0"
djoser = "^2.0.3"
psycopg2-binary = "^2.8.5"
python-decouple = "^3.3"
ipython = "^7.17.0"
ipdb = "^0.13.3"

[tool.poetry.dev-dependencies]
pytest = "^6.0.1"
pytest-django = "^3.9.0"
pytest-cov = "^2.10.0"
pylama = "^7.7.1"
flake8 = "^3.8.3"
isort = "^5.4.1"
black = "^19.10b0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
